feat(project): typescript opt-in
