+++ src/app/[lang]/(sitemap)/decouvrir/plan-site/page.js
@@ -1,15 +1,17 @@
 import 'server-only'
-import { fetchI18n, fetchPages, Sitemap } from '../../discover/sitemap/page'
+import { fetchI18n, fetchPages, Sitemap } from '../../utils'
 
 export async function generateStaticParams() {
-  return [{
-    lang: 'fr'
-  }]
+  return [
+    {
+      lang: 'fr',
+    },
+  ]
 }
 
 export default async function Page(props) {
-  const params = await props.params;
-  const i18n = await fetchI18n({ lang: params.lang})
+  const params = await props.params
+  const i18n = await fetchI18n({ lang: params.lang })
   const pages = await fetchPages({ lang: params.lang })
   return (
     <div className="prose dark:prose-invert max-w-none">
